requirements.txt
gunicorn
requests==2.25.1
numpy==1.16.6
scipy==1.2.3
django==3.2.4
python-decouple==3.4
djangorestframework==3.12.2
markdown==3.3.3
django-filter==2.4.0
django-tables2==2.4.0
user-agents==2.2.0
django-autocomplete-light==3.8.2
django-user-agents
djangorestframework-csv==2.1.1
